Configuration option error fixed
DEVICE_xxx was ENABLE_xxx. This was changed a while ago. How the old names
got back is a mystery, but since finding the error using git bisect didn't
work, the error must be part of each current commit and entered during
rebase which was done recently. I.e. automerge did it and the exact commit
introducing the error is "lost" as the change that should had been there
appearers to have never happened.

This was a tricky one as normally (i.e. in most cases, but also the most
probable) when automerge fails, the error will either be caught by
compilation or otherwise obvious. In this case it wasn't obvious
(link-error). Writing this down to have as a note about what can go wrong
using git.