Exclude workbench from git changes monitoring
Workbench is a git submodule that's a standalone program and that's supposed to
be changed easily. Therefore don't monitor this directory/sub-module for changes.

Note: It's unclear if this technique works as intended. This commit may need to
be reverted and replaced by something else.

Please see the following for techniques: http://goo.gl/fjKlsH