Permit cascading of options to workbench
Workbench can now parse it's own argument's. Everything after '--' is
sent to embedded-workbench's main(int argc, char **argv) in argc and argv
as if it would be when run without ehwe.

Note that not all, probably most, embedded targets will provide anything
meaningful in argc & argv (because it requires an environment, which
implies a process).