CS config messup fixed
There is no option to set CP polarity, but two ways to set CS.
* Via command: 0000001x ­ CS high (1) or low (0)
* Via peripheral configuration: Cmd 0100wxyz, bit z

Electric level follows logic level.

This could be a bug in FW as it would make sense only one method is needed
to alter CS and that polarity for CS active would make sense to have as it's
included in console mode.

This patch applies to Firmware: v5.10 (r559)