RT issue during state-change and timing.
State-change to I2C was prevented by a reply from a previous command which
was still pending. The same error should have affected previously finished
modality (SPI), but for happenstance reason didn't. The retry-loop handled
it for SPI, but is actually a workaround that hides real problem.

This made this bug hard to find. Retry-loop should be considered for
removal.

Solution:
* Added a delay prior sending state-change command to allow any on-the-way
  messages still in progress to propagate to stdlib buffer
* Throw these away

Please read patch as more weaknesses exists, and suggestions of how to solve
them.

Change-Id: I336842f515157e1c8ef3ea86968cd438fe19a12f