Timing adjust/contents identification mode-change
Identified which parts (3) contribute to timing. BP response time is the
only one not possible to deduct but has to be measured. As most hosts can't
guarantee timing, constant was chosen such that state transition succeeded
20 consecutive times.

This patch should contribute to robustness.

NOT SQUASHED to point out bottle-necks for future reference.

Possible improvement: Use synchronous mode for all state-change as soon as
raw mode is entered instead of postponing until targeted mode is entered.

Change-Id: I5d21b9a591b145bd8a7648832e6113dcbf3bb638