Fundamental rip-up & re-try: driver structure
A common driver-structure for all bus-types is not feasible as "interfaces"
(i.e. the simulated low-level API:s) require more specific function-calls
and reusing the same for different types of drivers just don't make any
sense.

This is opposed to a high level OS, where it's the other way around: The API
for using drivers is the same regardless of type (or "role" in ehwe
nomenclature).

So far only bus-pirate has been in question. The same divergence may be
needed for added physical interfaces, in which case the virtual interfaces
will have to have detailed knowledge about the physical one. If so this may
be unfortunate as the project could scale badly.

This commit is a massive search-replace commit which unless mistakes are
made, should have no run-time impact. It's however needed to finalize i2c
or any other role.

Change-Id: I1da8670d882731fb5456a56ed69790ac41119ad8