Generic interface heightens abstraction level
The generic interface added is a interface that lifts the abstraction level
and removes the ugly looking API names for the simulated interfaces (one
currently). It also allows freedom from the limitations some of the
simulated API:s have, and it allows to break "standard" from some of the
features/conventions e.t.c. we don't like.

This new API is currently fully dependant of the stm32 interface. We can
either work against removing this dependency, or consider this the native
(or back-bone) interface: Even though it's theoretically speaking not needed
as the real work is done in drivers.

If we separate them however, there will be a lot of duplication work. If we
don't, there will be a dependency that can't be broken by configure
(currently un-configuring stm32 will do just that).

Wether or not to go one way or the other is TBD

Change-Id: Ic7bd1873c944eaabd34f7ccc8b9e3d83051d48d4