i2c_device: Module initialization (dummy)
Final linkage complains about missing symbols though symbols are
existing in interfaces.a file.

To solve this without this work-around, libembedded.a would need
a build-system dependency to libinterfaces.a which would require
CMakeLists.txt to have deep knowledge about ehwe, which is
undesireable.

The work-around, in spite of doung nothing, isn't butt-ugly
though.