LXI: I2C sending using high-level API
First successful send of 1 and 2 bytes. For testing more bytes,
a device in the work-bench willing to accept more without
NACK-ing is needed (INA233 used).

Linux API object if NACK-ing. NACK-ing however is a valid way to
let sender know that slave is unwilling to receive more.
Adjustment if possible in the return handling might be needed to
not ASSURE (i.e. crash).

Invocation example:
  ./ehwe -d i2c:1:lxi:master:/dev/i2c-6